+++ packages/gatsby/src/query/query-compiler.js
@@ -15,6 +15,11 @@ const {
   validate,
   print,
   visit,
+  visitWithTypeInfo,
+  TypeInfo,
+  isAbstractType,
+  isObjectType,
+  isInterfaceType,
   Kind,
   FragmentsOnCompositeTypesRule,
   KnownTypeNamesRule,
@@ -345,7 +350,7 @@ const processDefinitions = ({
       continue
     }
 
-    const document = {
+    let document = {
       kind: Kind.DOCUMENT,
       definitions: Array.from(usedFragments.values())
         .map(name => definitionsByName.get(name).def)
@@ -384,6 +389,8 @@ const processDefinitions = ({
       }
       continue
     }
+
+    document = addExtraFields(document, schema)
 
     const query = {
       name,
@@ -464,4 +471,73 @@ const determineUsedFragmentsForDefinition = (
     }
     return { usedFragments, missingFragments }
   }
+}
+
+/**
+ * Automatically add:
+ *   `__typename` field to abstract types (unions, interfaces)
+ *   `id` field to all object/interface types having an id
+ * TODO: Remove this in v3.0 as it is a legacy from Relay compiler
+ */
+const addExtraFields = (document, schema) => {
+  const typeInfo = new TypeInfo(schema)
+  const contextStack = []
+
+  const transformer = visitWithTypeInfo(typeInfo, {
+    enter: {
+      [Kind.SELECTION_SET]: node => {
+        // Entering selection set:
+        //   selection sets can be nested, so keeping their metadata stacked
+        contextStack.push({ hasTypename: false, hasId: false })
+      },
+      [Kind.FIELD]: node => {
+        // Entering a field of the current selection-set:
+        //   mark which fields already exist in this selection set to avoid duplicates
+        const context = contextStack[contextStack.length - 1]
+        if (node.name.value === `__typename`) {
+          context.hasTypename = true
+        }
+        if (node.name.value === `id`) {
+          context.hasId = true
+        }
+      },
+    },
+    leave: {
+      [Kind.SELECTION_SET]: node => {
+        // Modify the selection-set AST on leave (add extra fields unless they already exist)
+        const context = contextStack.pop()
+        const parentType = typeInfo.getParentType()
+        const extraFields = []
+
+        // Adding __typename to unions and interfaces (if required)
+        if (!context.hasTypename && isAbstractType(parentType)) {
+          extraFields.push({
+            kind: Kind.FIELD,
+            name: { kind: Kind.NAME, value: `__typename` },
+          })
+        }
+        if (
+          !context.hasId &&
+          (isObjectType(parentType) || isInterfaceType(parentType)) &&
+          hasIdField(parentType)
+        ) {
+          extraFields.push({
+            kind: Kind.FIELD,
+            name: { kind: Kind.NAME, value: `id` },
+          })
+        }
+        return extraFields.length > 0
+          ? { ...node, selections: [...extraFields, ...node.selections] }
+          : undefined
+      },
+    },
+  })
+
+  return visit(document, transformer)
+}
+
+const hasIdField = type => {
+  const idField = type.getFields()[`id`]
+  const fieldType = idField ? String(idField.type) : ``
+  return fieldType === `ID` || fieldType === `ID!`
 }