+++ packages/gatsby-source-lever/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.24](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-lever@2.1.23...gatsby-source-lever@2.1.24) (2020-01-23)
+
+### Features
+
+- **gatsby-source-lever:** use schema customization API ([#20069](https://github.com/gatsbyjs/gatsby/issues/20069)) ([#20521](https://github.com/gatsbyjs/gatsby/issues/20521)) ([d02adb1](https://github.com/gatsbyjs/gatsby/commit/d02adb1))
+
 ## [2.1.23](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-lever@2.1.22...gatsby-source-lever@2.1.23) (2020-01-09)
 
 **Note:** Version bump only for package gatsby-source-lever