+++ packages/gatsby-theme-blog/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.34](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.2.33...gatsby-theme-blog@1.2.34) (2020-01-27)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.2.33](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.2.32...gatsby-theme-blog@1.2.33) (2020-01-27)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.2.32](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.2.31...gatsby-theme-blog@1.2.32) (2020-01-24)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.2.31](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.2.30...gatsby-theme-blog@1.2.31) (2020-01-24)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.2.30](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.2.29...gatsby-theme-blog@1.2.30) (2020-01-23)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.2.29](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.2.28...gatsby-theme-blog@1.2.29) (2020-01-23)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.2.28](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.2.27...gatsby-theme-blog@1.2.28) (2020-01-21)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
 ## [1.2.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.2.26...gatsby-theme-blog@1.2.27) (2020-01-21)
 
 **Note:** Version bump only for package gatsby-theme-blog